refactor: use shared modal directly in top-up sections
