+++ src/components/ui/Modal.tsx
@@ -1,0 +1,67 @@
+"use client";
+
+import * as React from "react";
+import { AnimatePresence, motion } from "framer-motion";
+
+interface ModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  children: React.ReactNode;
+  containerClassName?: string;
+  ariaLabelledby?: string;
+  ariaDescribedby?: string;
+}
+
+export function Modal({
+  isOpen,
+  onClose,
+  children,
+  containerClassName = "",
+  ariaLabelledby,
+  ariaDescribedby,
+}: ModalProps) {
+  React.useEffect(() => {
+    document.body.style.overflow = isOpen ? "hidden" : "auto";
+    return () => {
+      document.body.style.overflow = "auto";
+    };
+  }, [isOpen]);
+
+  return (
+    <AnimatePresence>
+      {isOpen && (
+        <motion.div
+          className="fixed inset-0 z-50 flex items-center justify-center p-4"
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          exit={{ opacity: 0 }}
+          role="dialog"
+          aria-modal="true"
+          aria-labelledby={ariaLabelledby}
+          aria-describedby={ariaDescribedby}
+        >
+          <motion.div
+            className="absolute inset-0 bg-black/60 dark:bg-black/70 backdrop-blur-sm"
+            onClick={onClose}
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            exit={{ opacity: 0 }}
+            aria-hidden="true"
+          />
+          <motion.div
+            className={`relative z-10 w-full max-w-xl rounded-2xl border border-gray-300 dark:border-primary-500/20 bg-white dark:bg-gray-900 shadow-lg p-8 transition-colors ${containerClassName}`}
+            initial={{ scale: 0.9, opacity: 0 }}
+            animate={{ scale: 1, opacity: 1 }}
+            exit={{ scale: 0.9, opacity: 0 }}
+            transition={{ type: "spring", stiffness: 200, damping: 25 }}
+          >
+            {children}
+          </motion.div>
+        </motion.div>
+      )}
+    </AnimatePresence>
+  );
+}
+
+export default Modal;
+
